requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-appconf==1.0.4
django-cors-headers==3.7.0
django-imagekit==4.0.2
django-rest-knox==4.1.0
djangorestframework==3.12.4
idna==2.10
oauthlib==3.1.1
pilkit==2.0
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
